# versions/before-backend-bugfix-20260922-133652/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RescheduleRequestCreate": {
        "properties": {
          "new_slot_id": {
            "format": "uuid",
            "title": "New Slot Id",
            "type": "string"
          }
        },
        "required": [
          "new_slot_id"
        ],
        "title": "RescheduleRequestCreate",
        "type": "object"
      },
...
      "TokenExchangeRequest": {
        "properties": {
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "TokenExchangeRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VisitRequestDetailOut": {
        "properties": {
          "age": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "assigned_staff_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Staff Id"
          },
          "campus_key": {
            "title": "Campus Key",
            "type": "string"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "follow_up_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Up At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "parent_name": {
            "title": "Parent Name",
            "type": "string"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "preferred_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Time"
          },
          "questions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Questions"
          },
          "slot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VisitSlotBriefOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "slot_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "campus_key",
          "status",
          "parent_name",
          "phone",
          "age",
          "preferred_time",
          "questions",
          "slot_id",
          "assigned_staff_id",
          "confirmed_at",
          "cancelled_at",
          "follow_up_at",
          "created_at"
        ],
        "title": "VisitRequestDetailOut",
        "type": "object"
      },
...
      "VisitSlotBriefOut": {
        "description": "案件上要顯示的「參觀時間」。容量與已預約數是時段管理頁的事，\n這裡只回日期與起訖；後台明細與列表都直接顯示這一份，不必再打一次\n/admin/slots 去換算家長約在哪一天。",
        "properties": {
          "end_time": {
            "format": "time",
            "title": "End Time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "slot_date": {
            "format": "date",
            "title": "Slot Date",
            "type": "string"
          },
          "start_time": {
            "format": "time",
            "title": "Start Time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slot_date",
          "start_time",
          "end_time"
        ],
        "title": "VisitSlotBriefOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Ivy Website Admin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/website/v1/admin/reschedule-requests": {
      "get": {
        "operationId": "list_reschedule_requests_api_website_v1_admin_reschedule_requests_get",
        "parameters": [
          {
            "in": "query",
            "name": "campus_key",
            "required": true,
            "schema": {
              "title": "Campus Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "ivy_admin_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ivy Admin Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Reschedule Requests Api Website V1 Admin Reschedule Requests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reschedule Requests",
        "tags": [
          "parent-access"
        ]
      }
    },
    "/api/website/v1/admin/reschedule-requests/{request_id}/approve": {
      "post": {
        "operationId": "approve_reschedule_request_api_website_v1_admin_reschedule_requests__request_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "ivy_admin_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ivy Admin Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitRequestDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Reschedule Request",
        "tags": [
          "parent-access"
        ]
      }
    },
    "/api/website/v1/admin/reschedule-requests/{request_id}/reject": {
      "post": {
        "operationId": "reject_reschedule_request_api_website_v1_admin_reschedule_requests__request_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "ivy_admin_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ivy Admin Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reject Reschedule Request Api Website V1 Admin Reschedule Requests  Request Id  Reject Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Reschedule Request",
        "tags": [
          "parent-access"
        ]
      }
    },
...
    "/api/website/v1/admin/visit-requests/{visit_request_id}/access-link": {
      "post": {
        "operationId": "create_parent_access_link_api_website_v1_admin_visit_requests__visit_request_id__access_link_post",
        "parameters": [
          {
            "in": "path",
            "name": "visit_request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Visit Request Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "ivy_admin_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ivy Admin Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Parent Access Link Api Website V1 Admin Visit Requests  Visit Request Id  Access Link Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Parent Access Link",
        "tags": [
          "parent-access"
        ]
      }
    },
...
    "/api/website/v1/public/visit-manage/cancel": {
      "post": {
        "operationId": "parent_cancel_api_website_v1_public_visit_manage_cancel_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ivy_parent_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ivy Parent Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitRequestDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parent Cancel",
        "tags": [
          "parent-access"
        ]
      }
    },
    "/api/website/v1/public/visit-manage/exchange": {
      "post": {
        "operationId": "exchange_parent_token_api_website_v1_public_visit_manage_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitRequestDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exchange Parent Token",
        "tags": [
          "parent-access"
        ]
      }
    },
    "/api/website/v1/public/visit-manage/me": {
      "get": {
        "operationId": "get_own_visit_request_api_website_v1_public_visit_manage_me_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ivy_parent_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ivy Parent Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitRequestDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Own Visit Request",
        "tags": [
          "parent-access"
        ]
      }
    },
    "/api/website/v1/public/visit-manage/reschedule-request": {
      "post": {
        "description": "只建立待核准紀錄，原時段維持不變，直到園方在 admin 端核准。",
        "operationId": "parent_request_reschedule_api_website_v1_public_visit_manage_reschedule_request_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "ivy_parent_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ivy Parent Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Parent Request Reschedule Api Website V1 Public Visit Manage Reschedule Request Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parent Request Reschedule",
        "tags": [
          "parent-access"
        ]
      }
    },
...
  }
}